Remove redux files, refactor Profile components and folders
