+++ frontend/src/components/Profile/styles.ts
