--- frontend/src/redux/actions-creators/profilepage.ts
@@ -1,8 +1,0 @@
-import {
-  ToogleProfilePageAction,
-  ToogleProfilePageTypes
-} from 'redux/types/profilepage';
-
-export function toogleProfilePage(): ToogleProfilePageAction {
-  return { type: ToogleProfilePageTypes.TOOGLE_PROFILE_PAGE };
-}
