+++ frontend/src/redux/reducers/index.ts
@@ -1,12 +1,10 @@
 import { combineReducers } from 'redux';
 import { userReducer } from './userReducer';
 import { todoReducer } from './todoReducer';
-import { toogleProfilePage } from './profilePageReducer';
 
 export const rootReducer = combineReducers({
   user: userReducer,
-  todo: todoReducer,
-  toogleProfile: toogleProfilePage
+  todo: todoReducer
 });
 
 export type RootState = ReturnType<typeof rootReducer>;
